Mouse controls are working
