+++ Assets/New Terrain.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c49de86ae57c29468469d57b48a765f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
